feat(rag): 增加可选的Reranker支持以提升混合检索精度
- 在AgentAutoConfiguration中引入Reranker及其默认实现IdentityReranker
- 修改HybridRetriever支持Reranker进行结果精排，提升Top-K准确性
- 在HybridRetriever中增加带Reranker的构造函数和重排序逻辑
- 当无自定义Reranker时，默认使用恒等实现避免影响现有功能
- 优化混合检索日志，增加重排序后结果的输出信息
- 新增SSE事件模型及序列化工具，支持流式事件的标准化传输
- 修正Resilience4jCircuitBreakerAdapter中接口调用次数和平均耗时计算逻辑

+++ ydsz-common/ydsz-common-sentry/src/main/java/com/njydsz/common/sentry/config/MetricsAutoConfiguration.java
@@ -128,7 +128,7 @@ public class MetricsAutoConfiguration {
     public CircuitBreakerRegistry circuitBreakerRegistry(SentryProperties properties) {
         SentryProperties.CircuitBreakerConfig cb = properties.getMetrics().getCircuitBreaker();
         CircuitBreakerConfig config = CircuitBreakerConfig.custom()
-                .failureRateThreshold(cb.getFailureRateThreshold())
+                .failureRateThreshold((float) cb.getFailureRateThreshold())
                 .slidingWindowType(CircuitBreakerConfig.SlidingWindowType.TIME_BASED)
                 .slidingWindowSize(cb.getSlidingWindowSize())
                 .waitDurationInOpenState(java.time.Duration.ofSeconds(cb.getHalfOpenAfterSeconds()))
